requirements.txt
dnspython>=2.3.0
cryptography>=42.0.0